MAINT: avoid use of deprecated cp.float, cp.int, cp.bool (#29)
This PR silences a number of warnings when running the tests with the current CuPy development branch. As in NumPy, these types are deprecated and `float`, `int` or `bool` should be used instead.

Authors:
  - Gregory R. Lee (https://github.com/grlee77)

Approvers:
  - https://github.com/jakirkham

URL: https://github.com/rapidsai/cucim/pull/29

+++ python/cucim/src/cucim/skimage/measure/tests/test_ccomp.py
@@ -55,7 +55,7 @@ class TestConnectedComponents:
         assert_array_equal(label(self.x, background=9), self.labels_bg_9)
 
     def test_random(self):
-        x = (cp.random.rand(20, 30) * 5).astype(cp.int)
+        x = (cp.random.rand(20, 30) * 5).astype(int)
         labels = label(x)
 
         n = int(labels.max())
@@ -180,7 +180,7 @@ class TestConnectedComponents3d:
         assert self.x[0, 0, 2] == 2, "Data was modified!"
 
     def test_random(self):
-        x = (cp.random.rand(20, 30) * 5).astype(cp.int)
+        x = (cp.random.rand(20, 30) * 5).astype(int)
         labels = label(x)
 
         n = int(labels.max())
